--- src/test/java/com/cognizant/javapunditsinvoicify/company/CompanyEntityIT.java
@@ -1,60 +1,0 @@
-package com.cognizant.javapunditsinvoicify.company;
-
-import com.cognizant.javapunditsinvoicify.dto.AddressDto;
-import com.cognizant.javapunditsinvoicify.dto.CompanyDto;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import org.junit.jupiter.api.Test;
-import org.mockito.InjectMocks;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.http.MediaType;
-import org.springframework.test.web.servlet.MockMvc;
-import org.springframework.test.web.servlet.RequestBuilder;
-import org.springframework.transaction.annotation.Transactional;
-
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
-import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-
-@SpringBootTest
-@AutoConfigureMockMvc
-@Transactional
-public class CompanyEntityIT {
-
-    @InjectMocks
-    private ObjectMapper objectMapper;
-
-    @Autowired
-    private MockMvc mockMvc;
-
-    @Test
-    public void createCompanyTest() throws Exception{
-
-        CompanyDto CompanyDto = new CompanyDto();
-        CompanyDto.setName("Name");
-        CompanyDto.setContactName("Contact Name");
-        CompanyDto.setContactTile("Contact Title");
-        CompanyDto.setContactNumber(123456789);
-        CompanyDto.setInvoices("Invoices");
-
-        AddressDto AddressDto = new AddressDto();
-        AddressDto.setLine1("Address line 1");
-        AddressDto.setLine2("line 2");
-        AddressDto.setCity("City");
-        AddressDto.setState("XX");
-        AddressDto.setZipcode(12345);
-
-        CompanyDto.setAddress(AddressDto);
-
-        RequestBuilder rq = post("/company")
-                .content(objectMapper.writeValueAsString(CompanyDto))
-                .contentType(MediaType.APPLICATION_JSON);
-
-        mockMvc.perform(rq)
-                .andExpect(status().isCreated())
-                .andDo(print())
-        ;
-    }
-
-}
